requirements.txt
grpcio-tools>=1.48.2
grpcio-reflection>=1.48.2
pyyaml
numpy>=1.24.0; python_version >= "3.8"
numpy>=1.19.5; python_version < "3.8"
PyOpenGL>=3.1.6
PyQt5
matplotlib
dataclasses
jinja2
h5py>=3.1.0
scipy
typing_extensions>=4.4.0; python_version >= "3.8"
typing_extensions; python_version < "3.8"
opencv-python
numpy-stl
scikit-learn
Pillow
lark
jsonpath_ng
nidaqmx
vtk
wheel
toml
